+++ packages/cli/test/execute/execute.test.ts
@@ -8,7 +8,6 @@ import { ExecuteOptions } from '../../src/execute/command';
 import handler from '../../src/execute/handler';
 import { mockFs, resetMockFs } from '../util';
 
-// Why is this logging everywhere?
 const logger = createMockLogger(undefined, { level: 'none' });
 
 // These options are designed to minimise i/o
@@ -20,12 +19,8 @@ const defaultOptions = {
   repoDir: '/repo',
   path: '.',
   log: {
-    // TODO if I pass a mock logger into the handler, the handler then
-    // goes and creates a real logger and passes it to the runtinme, which
-    // causes all sorts of logging
-    // Why isn't that affecting other tests?
-    // Why do I need this special handling here?
-    default: 'none',
+    // This special flag even hides errors
+    default: 'none-really',
   } as any, // TODO some weird typing here
   logJson: true,
 } as Partial<ExecuteOptions>;
@@ -128,6 +123,104 @@ test.serial('run a workflow with state', async (t) => {
   const result = await handler(options, logger);
   t.is(result.data.count, 4);
 });
+
+test.serial(
+  'run a workflow with errors and positions with anonymous steps',
+  async (t) => {
+    const workflow = {
+      workflow: {
+        steps: [
+          {
+            expression: `${fn}fn((state) => state.x.length)`,
+          },
+        ],
+      },
+    };
+
+    mockFs({
+      '/workflow.json': JSON.stringify(workflow),
+    });
+
+    const options = {
+      ...defaultOptions,
+      workflowPath: '/workflow.json',
+    };
+    const result = await handler(options, logger);
+
+    t.truthy(result.errors);
+    const err = result.errors['job-1'];
+    t.regex(err.message, /typeerror: cannot read properties of undefined/i);
+    t.is(err.pos.line, 2);
+    t.is(err.pos.column, 23);
+  }
+);
+
+test.serial(
+  "run a workflow with errors and positions with id'd steps",
+  async (t) => {
+    const workflow = {
+      workflow: {
+        steps: [
+          {
+            id: 'x',
+            expression: `${fn}fn((state) => state.x.length)`,
+          },
+        ],
+      },
+    };
+
+    mockFs({
+      '/workflow.json': JSON.stringify(workflow),
+    });
+
+    const options = {
+      ...defaultOptions,
+      workflowPath: '/workflow.json',
+    };
+    const result = await handler(options, logger);
+    t.truthy(result.errors);
+    t.regex(
+      result.errors.x.message,
+      /typeerror: cannot read properties of undefined/i
+    );
+    t.is(result.errors.x.pos.line, 2);
+    t.is(result.errors.x.pos.column, 23);
+  }
+);
+
+test.serial(
+  'run a workflow with errors and positions and named steps',
+  async (t) => {
+    const workflow = {
+      workflow: {
+        steps: [
+          {
+            id: 'x',
+            name: 'My Step',
+            expression: `${fn}fn((state) => state.x.length)`,
+          },
+        ],
+      },
+    };
+
+    mockFs({
+      '/workflow.json': JSON.stringify(workflow),
+    });
+
+    const options = {
+      ...defaultOptions,
+      workflowPath: '/workflow.json',
+    };
+    const result = await handler(options, logger);
+    t.truthy(result.errors);
+    t.regex(
+      result.errors.x.message,
+      /typeerror: cannot read properties of undefined/i
+    );
+    t.is(result.errors.x.pos.line, 2);
+    t.is(result.errors.x.pos.column, 23);
+  }
+);
 
 test.serial('run a workflow with cached steps', async (t) => {
   const workflow = {